convert vertices and indices into object

+++ src/ereacore/Direction.h
@@ -3,9 +3,9 @@
 #include <glm/glm.hpp>
 
 namespace Direction {
-    constexpr glm::vec3 UP = glm::vec3(0,1,0);
+    constexpr glm::ivec3 UP = glm::vec3(0,1,0);
 
     constexpr int ALL_SIZE = 6;
-    const glm::vec3* ALL = new glm::vec3[ALL_SIZE]{UP,UP,UP,UP,UP,UP,};
-    const std::vector<glm::vec3> ALL_VEC = {ALL,ALL+ALL_SIZE};
+    const glm::ivec3* ALL = new glm::ivec3[ALL_SIZE]{UP,UP,UP,UP,UP,UP,};
+    const std::vector<glm::ivec3> ALL_VEC = {ALL,ALL+ALL_SIZE};
 }
